requirements.txt
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
colorama==0.4.6
fastapi-slim==0.111.0
h11==0.16.0
httptools==0.7.1
idna==3.11
pydantic==2.12.4
pydantic-settings==2.2.1
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
sniffio==1.3.1
starlette==0.37.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.32.1
watchfiles==1.1.1
websockets==15.0.1
